Updated v6 to move todo and tododb classes out of program.cs to be consistant
