+++ aspnetcore/tutorials/min-web-api/samples/6.x/todo/Program.cs
@@ -126,24 +126,5 @@ app.MapDelete("/todoitems/{id}", async (int id, TodoDb db) =>
 #endregion
 
 app.Run();
-
-#region snippet_model
-public class Todo
-{
-    public int Id { get; set; }
-    public string? Name { get; set; }
-    public bool IsComplete { get; set; }
-}
-#endregion
-
-#region snippet_cntx
-class TodoDb : DbContext
-{
-    public TodoDb(DbContextOptions<TodoDb> options)
-        : base(options) { }
-
-    public DbSet<Todo> Todos => Set<Todo>();
-}
-#endregion
 #endregion
 #endif
