+++ aspnetcore/tutorials/min-web-api/samples/6.x/todo/Todo.cs
@@ -1,0 +1,6 @@
+public class Todo
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public bool IsComplete { get; set; }
+}
