+++ aspnetcore/tutorials/min-web-api/samples/6.x/todo/TodoDb.cs
@@ -1,0 +1,9 @@
+using Microsoft.EntityFrameworkCore;
+
+class TodoDb : DbContext
+{
+    public TodoDb(DbContextOptions<TodoDb> options)
+        : base(options) { }
+
+    public DbSet<Todo> Todos => Set<Todo>();
+}
